Resolve Compiler Warnings (#64)

+++ Source/Core/Manager/ERS_SceneManager/ERS_SceneManager.cpp
@@ -46,7 +46,7 @@ bool ERS_CLASS_SceneManager::AddScene(ERS_STRUCT_Scene Scene) {
 void ERS_CLASS_SceneManager::Render(ERS_STRUCT_OpenGLDefaults* OpenGLDefaults, std::shared_ptr<ERS_STRUCT_Shader> Shader) {
     
     // Iterate Through Models
-    for (long i = 0; i < (long)Scenes_[ActiveScene_]->Models.size(); i++) {
+    for (long i = 0; (long)i < (long)Scenes_[ActiveScene_]->Models.size(); i++) {
 
         // Get Model Pointer
         ERS_STRUCT_Model *Model = Scenes_[ActiveScene_]->Models[i].get();
